feat: entity spec
